First commit of scripts used for annotation
Scripts for summarizing and handling blast output: retrieve_hits (calls
hits_pickler); print_queries. These need to be updated and integrated

blast_annotation writes tsv files from xml formatted blast output. It currently
only works for TCDB annotation, and needs to be expanded. Ideally it would be
part of a pipeline which writes to a sqlite database.

Annotate is intended to hold a class which handles the backend of annotation,
such as separating out sequences based on their annotation. Currently only
works for TCDB