Fixed bugs in Annotate and fasta_channel_families
Annotate was modified to only write tsv files that had information in them,
and fasta_channe_families was modified to support this change. This required
loading up a dictionary of the information parsed in from the source
tsv annotation file instead of writing as a stream, so it uses more memory.
This dictionary could maybe be pickled for later use.