Edited remove_splices to handle Matz lab style transcriptomes
Wrote a new function longest_variants_matz. This uses a slightly different
strategy than the others in that it builds a dictionary from all genes found
in the whole file before flushing out longest variants. This has the advantage
of not needing the genes to be next to eachother in the file, but may have the
disadvantage of using more memory. Might be nice to work the flexibility part
into the other functions with an option.

Also edited get_longest by commenting out a useless piece of code. At least I
hope it's useless. Better double check by re-running on Ensembl data again.